+++ tests/__init__.py
@@ -1,0 +1,4 @@
+"""
+The `/tests` package contains all necessary files in order to support
+the execution of tests under `/ninecms/tests`
+"""